telegram_bot/requirements.txt
```pip-requirements
# Requirements for running the Telegram bookmark bot
requests
beautifulsoup4
pyrogram
python-dotenv
```